Break command prompt into two components
