+++ src/Assets/fonts.sass
@@ -15,6 +15,7 @@ body
     --size-logo: 22px
     --size-title: 22px
     --size-subtitle: 17px
+    --size-note: 14px
 
 $family-normal: var(--font-family-normal)
 $family-logo: var(--font-family-logo)
@@ -27,3 +28,4 @@ $size-normal: var(--size-normal)
 $size-logo: var(--size-logo)
 $size-title: var(--size-title)
 $size-subtitle: var(--size-subtitle)
+$size-note: var(--size-note)
