+++ src/Components/CommandItem/CommandItem.module.sass
@@ -1,0 +1,34 @@
+@use "../../Assets/colors"
+@use "../../Assets/dimensions"
+@use "../../Assets/fonts"
+
+.commandElement
+    padding: 0 16px
+    height: calc(28px + 2 * 16px)
+    font-size: calc(28px * 2 / 3)
+
+    &.selectedCommand
+        background-color: colors.$accent
+        color: colors.$background-primary
+
+        .matchedCommandPart
+            color: colors.$background-secondary !important
+            border-bottom: solid 2px colors.$background-secondary
+
+    &:hover:not(.selectedCommand)
+        background-color: colors.$background-secondary
+
+    .commandInfos
+        display: block
+        margin-left: 12px
+
+        .commandName
+            display: inline-block
+
+            .matchedCommandPart
+                color: colors.$accent
+                font-weight: fonts.$weight-bold
+
+        .commandTooltip
+            font-size: fonts.$size-note
+            text-overflow: ellipsis
